% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Construct a reproducible machine-learning analysis}
\usage{
analysis(
  problem,
  dir,
  manifest = NULL,
  create = TRUE,
  overwrite_manifest = FALSE
)
}
\arguments{
\item{problem}{A problem object.}

\item{dir}{Character scalar. Analysis output directory.}

\item{manifest}{Optional manifest list. If NULL, an existing manifest will be
read if present, otherwise a default manifest will be created.}

\item{create}{Logical. Should the analysis directory be created?}

\item{overwrite_manifest}{Logical. Should an existing manifest be
overwritten?}
}
\value{
An object of class c("<problem_type>_analysis", "analysis").
}
\description{
Creates an analysis object linking a problem object to an output directory
and a model manifest.
}
\details{
The manifest is intentionally analysis agnostic. It contains only:
file_version, enabled models, relative result paths, and model parameters.
}
