% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{analysis_enabled_models}
\alias{analysis_enabled_models}
\title{Return enabled model names for an analysis}
\usage{
analysis_enabled_models(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
Character vector.
}
\description{
Return enabled model names for an analysis
}
