% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_status.R
\name{analysis_existing_models}
\alias{analysis_existing_models}
\title{Return enabled models with existing result folders}
\usage{
analysis_existing_models(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
Character vector of model names.
}
\description{
Return enabled models with existing result folders
}
