% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{analysis_manifest}
\alias{analysis_manifest}
\title{Return the manifest for an analysis}
\usage{
analysis_manifest(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
A model manifest list.
}
\description{
Return the manifest for an analysis
}
