% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{analysis_params}
\alias{analysis_params}
\title{Return model params for an analysis}
\usage{
analysis_params(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
Named list.
}
\description{
Return model params for an analysis
}
