% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{analysis_resultpaths}
\alias{analysis_resultpaths}
\title{Return result paths for an analysis}
\usage{
analysis_resultpaths(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
Named character vector.
}
\description{
Return result paths for an analysis
}
