% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{as_rsample.problem}
\alias{as_rsample.problem}
\title{Convert a problem-class object to rsample format}
\usage{
as_rsample.problem(x, ...)
}
\arguments{
\item{x}{a \link{problem} object}

\item{...}{unused}
}
\value{
a \link[rsample:rsample-package]{rsample::rsample-package} rset object
}
\description{
Convert a problem-class object to rsample format
}
