% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-brier_skillscore.R
\name{bss_class}
\alias{bss_class}
\alias{bss_class.data.frame}
\alias{bss_class_vec}
\title{Brier Skill Score for probability output of classification models}
\usage{
bss_class(data, ...)

\method{bss_class}{data.frame}(
  data,
  truth,
  ...,
  na_rm = TRUE,
  event_level = .event_level(),
  case_weights = NULL
)

bss_class_vec(
  truth,
  estimate,
  na_rm = TRUE,
  event_level = .event_level(),
  case_weights = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} containing the columns specified by \code{truth} and
\code{...}.}

\item{...}{A set of unquoted column names or one or more
\code{dplyr} selector functions to choose which variables contain the
class probabilities. If \code{truth} is binary, only 1 column should be selected,
and it should correspond to the value of \code{event_level}. Otherwise, there
should be as many columns as factor levels of \code{truth} and the ordering of
the columns should be the same as the factor levels of \code{truth}.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{na_rm}{A \code{logical} value indicating whether \code{NA}
values should be stripped before the computation proceeds.}

\item{event_level}{A single string. Either \code{"first"} or \code{"second"} to specify
which level of \code{truth} to consider as the "event". This argument is only
applicable when \code{estimator = "binary"}. The default uses an internal helper
that defaults to \code{"first"}.}

\item{case_weights}{The optional column identifier for case weights.
This should be an unquoted column name that evaluates to a numeric column
in \code{data}. For \verb{_vec()} functions, a numeric vector,
\code{\link[hardhat:importance_weights]{hardhat::importance_weights()}}, or \code{\link[hardhat:frequency_weights]{hardhat::frequency_weights()}}.}

\item{estimate}{If \code{truth} is binary, a numeric vector of class probabilities
corresponding to the "relevant" class. Otherwise, a matrix with as many
columns as factor levels of \code{truth}. \emph{It is assumed that these are in the
same order as the levels of \code{truth}.}}
}
\description{
Compute the Brier Skill Score for a classification model.
}
\details{
Brier score is a metric that should be minimised.
The output ranges from 0 to 1, with 1 indicating perfect predictions.

The Brier Skill Score is analogous to \link[yardstick:rsq_trad]{rsq_trad}
in regression models, but calculated from the brier scores of classification
model predicted probabilities. The \link[yardstick:brier_class]{Brier score} is
the difference between a binary indicator for a class and its corresponding
class probability, squared and averaged. To convert to a skill score the
formula:

\deqn{BSS = 1 - \frac{BS_{\text{model}}}{BS_{\text{ref}}}}

Smaller values of the score are associated with better model performance.
}
\seealso{
\link[yardstick:brier_class]{brier_class},
\link[yardstick:rsq_trad]{rsq_trad},
\link[yardstick:prob-metrics]{All yardstick probability metrics}
}
\concept{class probability metrics}
