% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_fisherz}
\alias{combine_fisherz}
\title{Fisher-z combiner}
\usage{
combine_fisherz(x, ...)
}
\arguments{
\item{x}{Numeric vector of correlations.}

\item{...}{Additional arguments passed to mean.}
}
\value{
Numeric scalar.
}
\description{
Fisher-z combiner
}
