% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_mean}
\alias{combine_mean}
\title{Arithmetic mean combiner}
\usage{
combine_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical scalar.}
}
\value{
Numeric scalar.
}
\description{
Arithmetic mean combiner
}
