% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{combine_transformed}
\alias{combine_transformed}
\title{Transformed-scale combiner}
\usage{
combine_transformed(transform, inverse, summary = mean, ...)
}
\arguments{
\item{transform}{Function applied before summarising.}

\item{inverse}{Function applied after summarising.}

\item{summary}{Function used to summarise transformed values.}

\item{...}{passed to summary}
}
\value{
A function.
}
\description{
Transformed-scale combiner
}
