% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_analysis_in_dir}
\alias{create_analysis_in_dir}
\title{Create an analysis in a directory}
\usage{
create_analysis_in_dir(
  problem,
  dir,
  problem_path,
  manifest_path,
  manifest = NULL,
  create = TRUE,
  overwrite_problem = FALSE,
  overwrite_manifest = FALSE
)
}
\arguments{
\item{problem}{A problem object.}

\item{dir}{Character scalar.}

\item{problem_path}{Character scalar.}

\item{manifest_path}{Character scalar.}

\item{manifest}{Optional manifest list.}

\item{create}{Logical.}

\item{overwrite_problem}{Logical.}

\item{overwrite_manifest}{Logical.}
}
\value{
An analysis object.
}
\description{
Create an analysis in a directory
}
\keyword{internal}
