% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmultipredict-package.R
\docType{package}
\name{cvmultipredict-package}
\alias{cvmultipredict-package}
\title{cvmultipredict: Cross-validated prediction utilities}
\description{
Tools for fitting predictive models under cross-validation
and generating out-of-sample predictions.
}
\section{Main functions}{

\itemize{
\item \link{problem} - define a \code{problem} from an outcome vector,
a predictor matrix, and a cross-validation resampling scheme
\item \link{manifest} - JSON specification of prediction models and their settings
\item \link{analysis} - link a \code{problem} to a folder, and \code{manifest} of
prediction models to fit
\item \link{cvmultipredict} - run prediction models associated with an \code{analysis}
\item \link{pool_results} - collate cross-validated performance metrics from
\code{cvmultipredict}
\item \link{plot_pooled_results} - plot performance metrics
\item \link{plot_tuning_results} - plot hyperparameter tuning (for a given model)
}
}

\author{
\strong{Maintainer}: Andrew J. Lawrence \email{lawrenceajl@gmail.com}

Authors:
\itemize{
  \item Andrew J. Lawrence \email{lawrenceajl@gmail.com}
}

}
