% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmultipredict-package.R
\docType{package}
\name{cvmultipredict}
\alias{cvmultipredict-package}
\alias{cvmultipredict}
\title{cvmultipredict: Cross-validated prediction utilities}
\description{
Tools for fitting predictive models under cross-validation
and generating out-of-sample predictions.
}
\section{Main functions}{

TODO: information here.
}

\author{
\strong{Maintainer}: Andrew J. Lawrence \email{lawrenceajl@gmail.com}

Authors:
\itemize{
  \item Andrew J. Lawrence \email{lawrenceajl@gmail.com}
}

}
