% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_elasticnet.R
\name{elasticnet}
\alias{elasticnet}
\alias{elasticnet.default}
\alias{elasticnet.regression_analysis}
\alias{elasticnet.classification_analysis}
\title{A elastic-net regularised linear model with grid-tuned regularisation.}
\usage{
elasticnet(
  x,
  folder,
  penalty_res = 1000L,
  mixture_res = 3L,
  penalty_minratio = "auto",
  mixture_min = log(0.05),
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  ...
)

\method{elasticnet}{default}(
  x,
  folder,
  penalty_res = 1000L,
  mixture_res = 3L,
  penalty_minratio = "auto",
  mixture_min = log(0.05),
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  ...
)

\method{elasticnet}{regression_analysis}(
  x,
  folder,
  penalty_res = 1000L,
  mixture_res = 3L,
  penalty_minratio = "auto",
  mixture_min = 0.05,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = metricset_regression(),
  ...
)

\method{elasticnet}{classification_analysis}(
  x,
  folder,
  penalty_res = 1000L,
  mixture_res = 3L,
  penalty_minratio = "auto",
  mixture_min = 0.05,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = metricset_classification(),
  ...
)
}
\arguments{
\item{x}{an analysis object to run elasticnet models for.}

\item{folder}{a folder path ending in a model-name unique label
used to store results.}

\item{penalty_res}{the number of penalty values to tune over, values are
determined from the data using the glmnet method.}

\item{mixture_res}{the number of mixtures to tune over - values will be
log spaced between \code{mixture_min} and 1.0.}

\item{penalty_minratio}{see \code{lambda.min.ratio} argument in glmnet.
Set this to auto to follow glmnet's default behaviour.}

\item{mixture_min}{the smallest mixture value to consider - this is how close
the model should get to a pure L2 (Ridge) model. For p >> n the pure ridge
is very time-consuming, so the default is 0.05.
Set this to 0.0 to consider the time-consuming pure L2 Ridge when tuning.}

\item{tuning_resample_fxn}{a string containing a function name from \link[rsample:rsample]{rsample}
which will be used to resample the data for tuning purposes
(e.g. \link[rsample:bootstraps]{"bootstraps"}, \link[rsample:vfold_cv]{"vfoldcv"})}

\item{tuning_resample_args}{list of arguments to pass to \code{tuning_resample_fxn}}

\item{check_futility}{logical. Futility checks on the finalmodel
determine whether cross-validation should be conducted.
If \code{check_futility = FALSE} then cross-validation is
always conducted.}

\item{...}{unused.}

\item{metricset}{used internally to specify metrics}
}
\description{
Fits to data a model linear in the predictors but with regularisation to
reduce overfitting and perform variable selection (L1 penalty-type).
Tunes hyperparameters for:
\enumerate{
\item \strong{penalty} level (i.e. \link[glmnet:glmnet]{lambda}), and
\item the \strong{mixture} of L2-L1 penalty types (i.e. \link[glmnet:glmnet]{alpha})
}
}
\details{
Default tuning considers:
\itemize{
\item 3 mixture values of alpha with log-spacing between 0.05 and 1.0, so:
0.05, ~0.224, 1.0
\item 1000 penalty values of lambda with log10-spacing between the (data-driven)
maximum lambda (using glmnet to estimate this), and a minimum lambda derived
from a ratio parameter (\code{penalty_minratio}) multiplied by the maximum lambda.
}
}
\note{
Tuning uses a custom grid because fitting an elastic-net model produces
all values of the penalty parameter with minimal overhead (due to warm starts).
As a result, tuning over a large number of values of the regularisation penalty
(lambda) is computationally cheap, while tuning lots of values of the mixture
hyperparameter (alpha) is expensive (and often in practice less impactful).
Furthermore, the maximum lambda depends on the value of alpha selected.
}
\seealso{
\link[glmnet:glmnet]{glmnet::glmnet} \link[parsnip:linear_reg]{parsnip::linear_reg} \link[parsnip:logistic_reg]{parsnip::logistic_reg}
}
