% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{enabled_models}
\alias{enabled_models}
\alias{enabled_models.analysis}
\alias{enabled_models.default}
\title{Return enabled model names from a manifest}
\usage{
enabled_models(x, ...)

\method{enabled_models}{analysis}(x, ...)

\method{enabled_models}{default}(x, ...)
}
\arguments{
\item{x}{A model manifest}

\item{...}{unused}
}
\value{
Character vector.
}
\description{
Return enabled model names from a manifest
}
