% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{existing_models}
\alias{existing_models}
\alias{existing_models.default}
\alias{existing_models.analysis}
\title{Return enabled models with existing result folders}
\usage{
existing_models(x, ...)

\method{existing_models}{default}(x, ...)

\method{existing_models}{analysis}(x, ...)
}
\arguments{
\item{x}{An analysis object.}

\item{...}{unused.}
}
\value{
Character vector of existing model names.
}
\description{
Intended to support a future run.analysis() method.
}
