% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{foldernames}
\alias{foldernames}
\alias{foldernames.analysis}
\alias{foldernames.default}
\title{Return model folder paths from a manifest}
\usage{
foldernames(x, ...)

\method{foldernames}{analysis}(x, ...)

\method{foldernames}{default}(x, ...)
}
\arguments{
\item{x}{An analysis or its model manifest}

\item{...}{unused}
}
\value{
Named character vector.
}
\description{
Return model folder paths from a manifest
}
