% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{function_names}
\alias{function_names}
\alias{function_names.analysis}
\alias{function_names.default}
\title{Return function_name vector from a manifest}
\usage{
function_names(x, ...)

\method{function_names}{analysis}(x, ...)

\method{function_names}{default}(x, ...)
}
\arguments{
\item{x}{An analysis or its model manifest}

\item{...}{unused}
}
\value{
Named character vector.
}
\description{
Return function_name vector from a manifest
}
