% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{has_parent_path_component}
\alias{has_parent_path_component}
\title{Check whether a relative path contains parent directory components}
\usage{
has_parent_path_component(path)
}
\arguments{
\item{path}{Character scalar.}
}
\value{
Logical scalar.
}
\description{
Check whether a relative path contains parent directory components
}
\keyword{internal}
