% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{Check whether a path is absolute}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{Character scalar.}
}
\value{
Logical scalar.
}
\description{
Check whether a path is absolute
}
\keyword{internal}
