% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest}
\alias{manifest}
\title{Read an analysis manifest}
\usage{
manifest(x, ...)
}
\arguments{
\item{x}{Object to read manifest for.}

\item{...}{Additional arguments.}
}
\value{
a list setting out file_version and models: a list of
the models to run for a given analysis.
}
\description{
The manifest of an analysis is a JSON specifying which models to run, along
with additional options.
}
