% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{manifest_params}
\alias{manifest_params}
\title{Return model params from a manifest}
\usage{
manifest_params(manifest)
}
\arguments{
\item{manifest}{A model manifest list.}
}
\value{
Named list.
}
\description{
Return model params from a manifest
}
