% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{manifest_resultpaths}
\alias{manifest_resultpaths}
\title{Return model result paths from a manifest}
\usage{
manifest_resultpaths(manifest)
}
\arguments{
\item{manifest}{A model manifest list.}
}
\value{
Named character vector.
}
\description{
Return model result paths from a manifest
}
