% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{new_problem}
\alias{new_problem}
\title{Low-level problem constructor}
\usage{
new_problem(label, ytype, y, x, ids, folds)
}
\arguments{
\item{label}{Character scalar identifying the problem.}

\item{ytype}{Character scalar giving the response type:
"regression" or "classification".}

\item{y}{Response vector.}

\item{x}{Data frame of predictors.}

\item{ids}{Character vector of row identifiers.}

\item{folds}{List of training indices.}
}
\value{
An object with class c(ytype, "problem").
}
\description{
Low-level problem constructor
}
\keyword{internal}
