% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{null_model_mse}
\alias{null_model_mse}
\title{Null model performance}
\usage{
null_model_mse(x, ...)
}
\arguments{
\item{x}{A problem object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Numeric scalar.
}
\description{
Calculates the performance of a null model for a problem.
}
\details{
For regression, the null model predicts the mean of y for every observation
and returns the mean squared error.

For classification, the null model predicts the majority class for every
observation and returns the 0-1 error rate.
}
