% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_null.R
\name{nullmetrics}
\alias{nullmetrics}
\title{nullmetrics}
\usage{
nullmetrics(
  x,
  metric_set = NULL,
  na.rm = FALSE,
  event_level = "second",
  label = "null"
)
}
\arguments{
\item{x}{a vector, either numeric (regression) or factor (classification)}

\item{metric_set}{a \link[yardstick:metric_set]{metric_set} to apply to x.
Note: default (NULL) uses the default sets associated with
\link{cvmultipredict}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}

\item{event_level}{(for classification) which level of x is the target?
"first" | "second"}

\item{label}{sets contents of label column in the returned results}
}
\value{
A tibble containing \code{label}, \code{.metric}, \code{.estimator} and
\code{.estimate} columns.
}
\description{
Produces the apparent performance metrics obtained from an
"intercept-only" null model for a numeric (regression) or
factor (classification) vector.
}
