% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{pending_models}
\alias{pending_models}
\alias{pending_models.default}
\alias{pending_models.analysis}
\title{Return enabled models with missing result folders}
\usage{
pending_models(x, ...)

\method{pending_models}{default}(x, ...)

\method{pending_models}{analysis}(x, ...)
}
\arguments{
\item{x}{An analysis object.}

\item{...}{unused.}
}
\value{
Character vector of pending model names.
}
\description{
Intended to support a future run.analysis() method.
}
