% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{print.analysis}
\alias{print.analysis}
\title{Print an analysis}
\usage{
\method{print}{analysis}(x, ...)
}
\arguments{
\item{x}{An analysis object.}

\item{...}{Ignored.}
}
\description{
Print an analysis
}
