% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{resultpaths}
\alias{resultpaths}
\alias{resultpaths.analysis}
\alias{resultpaths.default}
\title{Return model result paths from a manifest}
\usage{
resultpaths(x, ...)

\method{resultpaths}{analysis}(x, ...)

\method{resultpaths}{default}(x, ...)
}
\arguments{
\item{x}{An analysis or its model manifest}

\item{...}{unused}
}
\value{
Named character vector.
}
\description{
Return model result paths from a manifest
}
