% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{resume_analysis_from_dir}
\alias{resume_analysis_from_dir}
\title{Resume an analysis from a directory}
\usage{
resume_analysis_from_dir(dir, problem_path, manifest_path)
}
\arguments{
\item{dir}{Character scalar.}

\item{problem_path}{Character scalar.}

\item{manifest_path}{Character scalar.}
}
\value{
An analysis object.
}
\description{
Resume an analysis from a directory
}
\keyword{internal}
