% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectbyrfimp.R
\name{step_selectbyrfimp}
\alias{step_selectbyrfimp}
\title{Select predictors using random forest importance}
\usage{
step_selectbyrfimp(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  n_selected = 10L,
  outcome = NULL,
  selected_predictors = NULL,
  skip = FALSE,
  id = recipes::rand_id("selectbyrfimp")
)
}
\arguments{
\item{recipe}{A recipe object.}

\item{...}{Predictor selectors.}

\item{role}{Not used.}

\item{trained}{Logical.}

\item{n_selected}{Number of variables to retain.}

\item{outcome}{Name of the outcome variable.}

\item{selected_predictors}{Filled during prep().}

\item{skip}{Skip when baking new data.}

\item{id}{Unique id.}
}
\description{
Select predictors using random forest importance
}
