% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summary.analysis}
\alias{summary.analysis}
\title{Summarise an analysis}
\usage{
\method{summary}{analysis}(object, ...)
}
\arguments{
\item{object}{An analysis object.}

\item{...}{Ignored.}
}
\value{
An object of class "summary.analysis".
}
\description{
Summarise an analysis
}
