% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{summary.problem}
\alias{summary.problem}
\title{Summarise a problem}
\usage{
\method{summary}{problem}(object, ...)
}
\arguments{
\item{object}{A problem object.}

\item{...}{Ignored.}
}
\description{
Summarise a problem
}
