% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{validate_analysis}
\alias{validate_analysis}
\title{Validate an analysis object}
\usage{
validate_analysis(x)
}
\arguments{
\item{x}{An analysis object.}
}
\value{
The validated analysis object.
}
\description{
Validate an analysis object
}
\keyword{internal}
