% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{validate_model_manifest}
\alias{validate_model_manifest}
\title{Validate a model manifest}
\usage{
validate_model_manifest(manifest)
}
\arguments{
\item{manifest}{A model manifest list.}
}
\value{
The validated manifest.
}
\description{
Validate a model manifest
}
