% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmultipredict.R
\name{cvmultipredict}
\alias{cvmultipredict}
\title{Apply to an analysis to run models or collate results from disk}
\usage{
cvmultipredict(x, .RUN = TRUE)
}
\arguments{
\item{x}{an \link{analysis} object (which links a \link{problem} and a \link{manifest} to a
working directory)}

\item{.RUN}{should missing models be calculated?
If FALSE then cvmultipredict will attempt to load pre-calculated
results from disk. At least the final model stage must have completed.}
}
\value{
A named list containing, for each model in the \link{manifest},
a data.frame of performance metrics.
}
\description{
Apply to an analysis to run models or collate results from disk
}
\seealso{
\link{problem} \link{manifest} \link{analysis} \link{pool_results}
}
