% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmultipredict-package.R, R/cvmultipredict.R
\docType{package}
\name{cvmultipredict}
\alias{cvmultipredict-package}
\alias{cvmultipredict}
\title{cvmultipredict: Cross-validated prediction utilities}
\usage{
cvmultipredict(x)
}
\arguments{
\item{x}{an \link{analysis} object linking a \link{problem} and a \link{manifest} to a
working directory}
}
\value{
a list of results from the models which were run
}
\description{
Tools for fitting predictive models under cross-validation
and generating out-of-sample predictions.
}
\section{Main functions}{

TODO: information here.
}

\author{
\strong{Maintainer}: Andrew J. Lawrence \email{lawrenceajl@gmail.com}

Authors:
\itemize{
  \item Andrew J. Lawrence \email{lawrenceajl@gmail.com}
}

}
