% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{default_model_manifest}
\alias{default_model_manifest}
\title{Create a default model manifest}
\usage{
default_model_manifest()
}
\value{
A model manifest list.
}
\description{
Create a default model manifest
}
