% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{identical_problem_identity}
\alias{identical_problem_identity}
\title{Compare lightweight identity of two problem objects}
\usage{
identical_problem_identity(x, y)
}
\arguments{
\item{x}{A problem object.}

\item{y}{A problem object.}
}
\value{
Logical scalar.
}
\description{
This checks whether two problem objects appear to describe the same problem
without requiring full object identity.
}
\keyword{internal}
