% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_linear.R
\name{linear}
\alias{linear}
\alias{linear.default}
\alias{linear.regression_analysis}
\alias{linear.classification_analysis}
\title{A (general) linear model with minimal (untuned) regularisation.}
\usage{
linear(x, folder, pca = "auto", pca_ncomp = "auto", ...)

\method{linear}{default}(x, folder, pca = "auto", pca_ncomp = "auto", ...)

\method{linear}{regression_analysis}(
  x,
  folder,
  pca = "auto",
  pca_ncomp = "auto",
  model = parsnip::linear_reg(),
  metricset = metricset_regression(),
  ...
)

\method{linear}{classification_analysis}(
  x,
  folder,
  pca = "auto",
  pca_ncomp = "auto",
  model = parsnip::logistic_reg(),
  metricset = metricset_classification(),
  ...
)
}
\arguments{
\item{x}{an analysis object to run linear models for.}

\item{folder}{a folder path ending in a model-name unique label
used to store results.}

\item{pca}{either TRUE (uses PCA to reduce input dimensionality),
FALSE (does not reduce dimensionality) or default: \code{"auto"}
such that auto applies a PCA step if n < (50 + 10p)}

\item{pca_ncomp}{number of PCA components, either positive integer <= p
or "auto" to use \code{floor((n - 50)/10)}.}

\item{...}{unused.}

\item{model}{used internally to specify a parsnip model}

\item{metricset}{used internally to specify metrics}
}
\description{
Either no regularisation is applied, or an untuned PCA reduction of the
predictors is used depending on settings for \code{pca} and \code{pca_ncomp}
}
\note{
For pca = "auto" calculations,
n is number of samples, p is number of predictors.
}
