% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_helpers.R
\name{manifest_params}
\alias{manifest_params}
\alias{manifest_params.analysis}
\alias{manifest_params.default}
\title{Return model setting params from a manifest}
\usage{
manifest_params(x, ...)

\method{manifest_params}{analysis}(x, ...)

\method{manifest_params}{default}(x, ...)
}
\arguments{
\item{x}{An analysis or its model manifest list.}

\item{...}{unused.}
}
\value{
Named list.
}
\description{
Return model setting params from a manifest
}
