% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{model_manifest_entry}
\alias{model_manifest_entry}
\title{Create a model entry for a model manifest}
\usage{
model_manifest_entry(function_name, resultpath, params = list())
}
\arguments{
\item{function_name}{the underlying model function used by cvmultipredict}

\item{resultpath}{Character scalar. Relative path from the analysis directory
to the model output directory.}

\item{params}{Named list. Model-specific fitting options.}
}
\value{
A list.
}
\description{
Create a model entry for a model manifest
}
\keyword{internal}
