% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_neuralnet.R
\name{neuralnet}
\alias{neuralnet}
\alias{neuralnet.default}
\alias{neuralnet.regression_analysis}
\alias{neuralnet.classification_analysis}
\title{This is a neuralnetwork model 2-layer multilayer perceptron}
\usage{
neuralnet(
  x,
  folder,
  tuning_grid_n = 243,
  tunevals_hidden_layers = 2^(2:6),
  tunevals_n_selected = seq(4, 12, by = 2),
  tunerange_epochs = c(10L, 1000L),
  tunerange_learn_rate = c(-10, -1),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = NULL,
  .RUN = TRUE,
  ...
)

\method{neuralnet}{default}(
  x,
  folder,
  tuning_grid_n = 243,
  tunevals_hidden_layers = 2^(2:6),
  tunevals_n_selected = seq(4, 12, by = 2),
  tunerange_epochs = c(10L, 1000L),
  tunerange_learn_rate = c(-10, -1),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = NULL,
  .RUN = TRUE,
  ...
)

\method{neuralnet}{regression_analysis}(
  x,
  folder,
  tuning_grid_n = 243,
  tunevals_hidden_layers = 2^(2:6),
  tunevals_n_selected = seq(4, 12, by = 2),
  tunerange_epochs = c(10L, 1000L),
  tunerange_learn_rate = c(-10, -1),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = metricset_regression(),
  .RUN = TRUE,
  ...
)

\method{neuralnet}{classification_analysis}(
  x,
  folder,
  tuning_grid_n = 243,
  tunevals_hidden_layers = 2^(2:6),
  tunevals_n_selected = seq(4, 12, by = 2),
  tunerange_epochs = c(10L, 1000L),
  tunerange_learn_rate = c(-10, -1),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  tuning_resample_fxn = "vfold_cv",
  tuning_resample_args = list(v = 10, repeats = 1, strata = "y"),
  check_futility = TRUE,
  metricset = metricset_classification(),
  .RUN = TRUE,
  ...
)
}
\arguments{
\item{x}{an analysis object to run models for.}

\item{folder}{a folder path ending in a model-name unique label
used to store results.}

\item{tuning_grid_n}{size of initial \link[dials:grid_space_filling]{space-filling}
tuning grid. Default: 3^5 = 243.
Recommendation: base_spacing ^ dimensionality -
i.e. same number of elements as a hypercube with a (relatively)
low spacing resolution in each dimension. Due to the space_filling grid the
effective spacing within dimension is much greater than the base spacing.}

\item{tunevals_hidden_layers}{tuning values for hidden_layers.
default: 4,8,16,32,64.
\emph{Used for h1 and h2 crossed. s.t. tuning includes h1 > h2 and h2 > h1.}}

\item{tunevals_n_selected}{tuning values (discrete) for number of variables to
include (filtered by \link[=step_selectbyrfimp]{rf importance}).}

\item{tunerange_epochs}{tuning range for \link[dials:epochs]{epochs}.
default: 10 : 1000}

\item{tunerange_learn_rate}{tuning range for \link[dials:learn_rate]{learn_rate}.
default 1e-10 : 0.1}

\item{tuning_bayes_maxit}{maximum number of Bayesian optimisation iterations.
Default: 50L. Set to 0L to suppress bayesian optimisation.}

\item{tuning_bayes_minit}{number of iterations without improvement before
Bayesian optimisation is terminated.
Default: 10}

\item{tuning_resample_fxn}{a string containing a function name from \link[rsample:rsample]{rsample}
which will be used to resample the data for tuning purposes
(e.g. \link[rsample:bootstraps]{"bootstraps"}, \link[rsample:vfold_cv]{"vfoldcv"})}

\item{tuning_resample_args}{list of arguments to pass to \code{tuning_resample_fxn}}

\item{check_futility}{logical. Futility checks on the finalmodel
determine whether cross-validation should be conducted.
If \code{check_futility = FALSE} then cross-validation is
always conducted.}

\item{metricset}{used internally to specify the outer-loop evaluation metrics
(not tuning)}

\item{.RUN}{used internally to allow results collation without running
missing models.}

\item{...}{unused.}
}
\description{
This is a neuralnetwork model 2-layer multilayer perceptron
}
