% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{new_analysis}
\alias{new_analysis}
\title{Low-level analysis constructor}
\usage{
new_analysis(problem, dir, problem_path, manifest_path)
}
\arguments{
\item{problem}{A problem object.}

\item{dir}{Character scalar. Analysis output directory.}

\item{problem_path}{Character scalar. Path to problem.RData.}

\item{manifest_path}{Character scalar. Path to model_manifest.json.}
}
\value{
An object of class c("<problem_type>_analysis", "analysis").
}
\description{
Low-level analysis constructor
}
\keyword{internal}
