% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{null_model_mse}
\alias{null_model_mse}
\title{Null model performance}
\usage{
null_model_mse(x, validity = c("apparent", "cv"), combine = combine_mean, ...)
}
\arguments{
\item{x}{A problem object.}

\item{validity}{Either "apparent" or "cv".}

\item{combine}{Function used to combine split-level estimates.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Numeric scalar.
}
\description{
Calculates null model mean squared error using either apparent validity
or cross-validation folds stored in the problem object.
}
