% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_pool.R
\name{plot_pooled_results}
\alias{plot_pooled_results}
\title{plot_pooled_results}
\usage{
plot_pooled_results(x)
}
\arguments{
\item{x}{result of \link{pool_results}}
}
\value{
A \link[ggplot2:ggplot]{ggplot object}
}
\description{
a \link[ggplot2:ggplot]{ggplot} linking the Apparent and (mean)
cross-validated performance of each metric.
}
