% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_pool.R
\name{pool_results}
\alias{pool_results}
\title{pool_results}
\usage{
pool_results(x, na.rm = FALSE)
}
\arguments{
\item{x}{\link{cvmultipredict} result.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}
}
\value{
A tibble containing columns:
\itemize{
\item \code{model}: label for the model.
\item \code{cvtype}: either \code{cv} or \code{app} indicating if estimate is cross-validated
or apparent.
\item \code{.metric}: the \link[yardstick:metrics]{metric}.
\item \code{.estimate}: metric value (for CV averaged over resamples)
\item \code{.std_err}: SD of the metric value (\code{NA} for apparent performance)
}
}
\description{
Calculates average performance metrics over cross-validation resamples.
Input \code{x} should be the return of cvmultipredict.
i.e. a named list of results data.frames - one for each model.
}
