% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problem_type}
\alias{problem_type}
\title{Return the type of a problem}
\usage{
problem_type(x)
}
\arguments{
\item{x}{A problem object.}
}
\value{
\code{"regression"} or \code{"classification"}.
}
\description{
Return the type of a problem
}
