% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{read_analysis_problem}
\alias{read_analysis_problem}
\title{Read an analysis problem from disk}
\usage{
read_analysis_problem(path)
}
\arguments{
\item{path}{Path to problem.RData.}
}
\value{
A problem object.
}
\description{
Reads a problem object from an RData file. The file must contain an object
named \code{problem}.
}
\keyword{internal}
