% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectbyrfimp.R
\name{rf_n_selected}
\alias{rf_n_selected}
\title{rf_n_selected}
\usage{
rf_n_selected(range = c(1L, dials::unknown()), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\description{
Number of predictors selected by Random Forest importance used in
\link{step_selectbyrfimp}.
}
\details{
This parameter is integer. Upper bound should be determined by
the number of predictors in the model.
}
