% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{validate_problem}
\alias{validate_problem}
\title{Validate a problem object}
\usage{
validate_problem(x)
}
\arguments{
\item{x}{A problem object.}
}
\value{
The validated object, invisibly.
}
\description{
Internal validator for objects of class "problem".
}
\keyword{internal}
