% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{write_analysis_problem}
\alias{write_analysis_problem}
\title{Write an analysis problem to disk}
\usage{
write_analysis_problem(problem, path)
}
\arguments{
\item{problem}{A problem object.}

\item{path}{Output path, usually file.path(dir, "problem.RData").}
}
\value{
The path, invisibly.
}
\description{
Saves the problem object to an RData file using the object name \code{problem}.
}
\keyword{internal}
