% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_xgboost.R
\name{xgboost}
\alias{xgboost}
\alias{xgboost.default}
\alias{xgboost.regression_analysis}
\alias{xgboost.classification_analysis}
\title{XGBoost model with grid-tuned hyperparameters followed by Bayesian optimisation}
\usage{
xgboost(
  x,
  folder,
  tuning_grid_n = 243,
  tunerange_min_n = c(2L, 10L),
  tunerange_trees = c(1L, 2000L),
  tunerange_loss_reduction = c(-3L, 2L),
  tunerange_tree_depth = c(2L, 10L),
  tunerange_learn_rate = c(-5, 1.5),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  ...
)

\method{xgboost}{default}(
  x,
  folder,
  tuning_grid_n = 243,
  tunerange_min_n = c(2L, 10L),
  tunerange_trees = c(1L, 2000L),
  tunerange_loss_reduction = c(-3L, 2L),
  tunerange_tree_depth = c(2L, 10L),
  tunerange_learn_rate = c(-5, 1.5),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  ...
)

\method{xgboost}{regression_analysis}(
  x,
  folder,
  tuning_grid_n = 243,
  tunerange_min_n = c(2L, 10L),
  tunerange_trees = c(1L, 2000L),
  tunerange_loss_reduction = c(-3L, 2L),
  tunerange_tree_depth = c(2L, 10L),
  tunerange_learn_rate = c(-5, 1.5),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  metricset = metricset_regression(),
  tuning_resample_fxn = "bootstraps",
  tuning_resample_args = list(times = 25L, strata = "y"),
  ...
)

\method{xgboost}{classification_analysis}(
  x,
  folder,
  tuning_grid_n = 243,
  tunerange_min_n = c(2L, 10L),
  tunerange_trees = c(1L, 2000L),
  tunerange_loss_reduction = c(-3L, 2L),
  tunerange_tree_depth = c(2L, 10L),
  tunerange_learn_rate = c(-5, 1.5),
  tuning_bayes_maxit = 50L,
  tuning_bayes_minit = 10L,
  metricset = metricset_classification(),
  tuning_resample_fxn = "bootstraps",
  tuning_resample_args = list(times = 25L, strata = "y"),
  ...
)
}
\arguments{
\item{x}{an analysis object to run xgboost models for.}

\item{folder}{a folder path ending in a model-name unique label
used to store results.}

\item{tuning_grid_n}{size of initial \link[dials:grid_space_filling]{space-filling}
tuning grid. Default: 3^5 = 243.
Recommendation: base_spacing ^ dimensionality -
i.e. same number of elements as a hypercube with a (relatively)
low spacing resolution in each dimension. Due to the space_filling grid the
effective spacing within dimension is much greater than the base spacing.}

\item{tunerange_min_n}{tuning range for min_n. default = c(2L, 10L).}

\item{tunerange_trees}{tuning range for trees. default = c(1L, 2000L).}

\item{tunerange_loss_reduction}{tuning range for loss_reduction.
default = c(-3L, 2L). log10 scale.}

\item{tunerange_tree_depth}{tuning range for tree_depth. default = c(2L, 10L)}

\item{tunerange_learn_rate}{tuning range for learn_rate. default = c(-5, 1.5).
log10 scale.}

\item{tuning_bayes_maxit}{maximum number of Bayesian optimisation iterations.
Default: 50L. Set to 0L to suppress bayesian optimisation.}

\item{tuning_bayes_minit}{number of iterations without improvement before
Bayesian optimisation is terminated.
Default: 10}

\item{...}{unused.}

\item{metricset}{used internally to specify the outer-loop evaluation metrics
(not tuning)}

\item{tuning_resample_fxn}{a string containing a function name from \link[rsample:rsample]{rsample}
which will be used to resample the data for tuning purposes
(e.g. \link[rsample:bootstraps]{"bootstraps"}, \link[rsample:vfold_cv]{"vfoldcv"})}

\item{tuning_resample_args}{list of arguments to pass to \code{tuning_resample_fxn}}
}
\description{
Fits Extreme Gradient Boosting (XGBoost) models to the problem.
An XGBoost model is a non-linear ensemble of decision trees which are
built sequentially to improve upon the errors made in the previous trees.
Decision trees can automatically capture non-linear effects and higher order
interactions between features without requiring these be specified in
advance. Uses the \code{\link[xgboost:xgboost]{xgboost}} engine with tuning wrapped by
\link[parsnip:boost_tree]{parsnip}.
}
\note{
Hyperparameter ranges are length 2 or length 1 vectors. Use length 1
vectors to suppress tuning for that hyperparameter.
If a length 1 vector with value NA is provided then the
underlying default will be used (passing null)
Length 2 vectors are assumed to specify a range of hyperparameter
values to explore.
}
\section{Default Behaviour}{

Tunes the following over a \link[dials:grid_space_filling]{space filling grid}
with resolution^dimensionality = 3^5 = 243 entries:
\itemize{
\item ntrees - number of trees in the ensemble
\item tree_depth - number of splits
\item learn_rate - (i.e. eta) step-size shrinkage used to prevent overfitting
\item min_n - threshold number of observations in a potential leaf node for
further partitioning to be considered. Larger == more conservative.
\item loss_reduction - threshold minimum reduction in the loss function required
to make a further partition on a leaf node. Larger == more conservative.
}

After initial grid tuning, \link[tune:tune_bayes]{Bayesian optimisation}
(Gaussian Process model) is applied to fine tune for 10-50 iterations.

The following hyperparameters are fixed / untuned:
\itemize{
\item \code{mtry = 1.0} - do not randomly subsample features.
\item \code{sample_size = 1.0} - do not randomly subsample observations.
}

All other values remain at the defaults of \link[xgboost:xgboost]{xgboost::xgboost}
via \link[parsnip:xgb_train]{parsnip::xgb_train}
}

\seealso{
\link[xgboost:xgboost]{xgboost::xgboost} \link[parsnip:boost_tree]{parsnip::boost_tree}
}
